implement review suggestions
* move logic from src/binary/mod.rs to src/config.ts
* remove redundant clap alias
* add example of local_proxy.json in REDME.md
* remove the "wrong" comments from binary/proxy.rs
* give refresh_config_notify to signal hook task
